requirement.txt
langchain_pinecone==0.0.3
streamlit
langchain
transformers
PyPDF2
python-dotenv
sentence_transformers
chromadb
streamlit_option_menu
langchain-community
langchain-huggingface
selenium
beautifulsoup4
langchain-groq